all: Update flynn/pq.
