+++ Godeps/_workspace/src/github.com/flynn/pq/bench_test.go
@@ -36,7 +36,6 @@ func BenchmarkSelectSeries(b *testing.B) {
 }
 
 func benchQuery(b *testing.B, query string, result interface{}) {
-	b.Skip("current pq database-backed benchmarks are inconsistent")
 	b.StopTimer()
 	db := openTestConn(b)
 	defer db.Close()
@@ -184,7 +183,6 @@ func BenchmarkPreparedSelectSeries(b *testing.B) {
 }
 
 func benchPreparedQuery(b *testing.B, query string, result interface{}) {
-	b.Skip("current pq database-backed benchmarks are inconsistent")
 	b.StopTimer()
 	db := openTestConn(b)
 	defer db.Close()
