+++ Godeps/_workspace/src/github.com/flynn/pq/conn_test.go
@@ -5,8 +5,6 @@ import (
 	"io"
 	"os"
 	"reflect"
-	"runtime"
-	"strings"
 	"testing"
 	"time"
 
@@ -57,12 +55,6 @@ func getServerVersion(t *testing.T, db *sql.DB) int {
 }
 
 func TestReconnect(t *testing.T) {
-	if runtime.Version() == "go1.0.2" {
-		fmt.Println("Skipping failing test; " +
-			"fixed in database/sql on go1.0.3+")
-		return
-	}
-
 	db1 := openTestConn(t)
 	defer db1.Close()
 	tx, err := db1.Begin()
@@ -839,11 +831,6 @@ func TestIssue196(t *testing.T) {
 // Test that any CommandComplete messages sent before the query results are
 // ignored.
 func TestIssue282(t *testing.T) {
-	if strings.HasPrefix(runtime.Version(), "go1.0") {
-		fmt.Println("Skipping test due to lack of Queryer implementation")
-		return
-	}
-
 	db := openTestConn(t)
 	defer db.Close()
 
@@ -876,6 +863,60 @@ func TestReadFloatPrecision(t *testing.T) {
 	}
 	if float8val != float64(35.03554004971999) {
 		t.Errorf("Expected float8 fidelity to be maintained; got no match")
+	}
+}
+
+func TestXactMultiStmt(t *testing.T) {
+	// minified test case based on bug reports from
+	// pico303@gmail.com and rangelspam@gmail.com
+	t.Skip("Skipping failing test")
+	db := openTestConn(t)
+	defer db.Close()
+
+	tx, err := db.Begin()
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer tx.Commit()
+
+	rows, err := tx.Query("select 1")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if rows.Next() {
+		var val int32
+		if err = rows.Scan(&val); err != nil {
+			t.Fatal(err)
+		}
+	} else {
+		t.Fatal("Expected at least one row in first query in xact")
+	}
+
+	rows2, err := tx.Query("select 2")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if rows2.Next() {
+		var val2 int32
+		if err := rows2.Scan(&val2); err != nil {
+			t.Fatal(err)
+		}
+	} else {
+		t.Fatal("Expected at least one row in second query in xact")
+	}
+
+	if err = rows.Err(); err != nil {
+		t.Fatal(err)
+	}
+
+	if err = rows2.Err(); err != nil {
+		t.Fatal(err)
+	}
+
+	if err = tx.Commit(); err != nil {
+		t.Fatal(err)
 	}
 }
 
@@ -1169,14 +1210,15 @@ func TestRuntimeParameters(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		defer db.Close()
 
 		// application_name didn't exist before 9.0
 		if test.param == "application_name" && getServerVersion(t, db) < 90000 {
+			db.Close()
 			continue
 		}
 
 		tryGetParameterValue := func() (value string, outcome RuntimeTestResult) {
+			defer db.Close()
 			row := db.QueryRow("SELECT current_setting($1)", test.param)
 			err = row.Scan(&value)
 			if err != nil {
