+++ Godeps/_workspace/src/github.com/flynn/pq/copy.go
@@ -124,8 +124,6 @@ awaitCopyInResponse:
 			errorf("unknown response for CopyFail: %q", t)
 		}
 	}
-
-	panic("not reached")
 }
 
 func (ci *copyin) flush(buf []byte) {
