+++ Godeps/_workspace/src/github.com/flynn/pq/notify.go
@@ -170,8 +170,6 @@ func (l *ListenerConn) listenerConnLoop() (err error) {
 			return fmt.Errorf("unexpected message %q from server in listenerConnLoop", t)
 		}
 	}
-
-	panic("not reached")
 }
 
 // This is the main routine for the goroutine receiving on the database
@@ -318,8 +316,6 @@ func (l *ListenerConn) ExecSimpleQuery(q string) (executed bool, err error) {
 			return false, fmt.Errorf("unknown response for simple query: %q", m.typ)
 		}
 	}
-
-	panic("not reached")
 }
 
 func (l *ListenerConn) Close() error {
@@ -426,6 +422,13 @@ func NewListener(name string,
 	go l.listenerMain()
 
 	return l
+}
+
+// Returns the notification channel for this listener.  This is the same
+// channel as Notify, and will not be recreated during the life time of the
+// Listener.
+func (l *Listener) NotificationChannel() <-chan *Notification {
+	return l.Notify
 }
 
 // Listen starts listening for notifications on a channel.  Calls to this
@@ -631,8 +634,6 @@ func (l *Listener) resync(cn *ListenerConn, notificationChan <-chan *Notificatio
 			return err
 		}
 	}
-
-	panic("not reached")
 }
 
 // caller should NOT be holding l.lock
