+++ Godeps/_workspace/src/github.com/flynn/pq/notify_test.go
@@ -24,7 +24,6 @@ func expectNotification(t *testing.T, ch <-chan *Notification, relname string, e
 	case <-time.After(1500 * time.Millisecond):
 		return fmt.Errorf("timeout")
 	}
-	panic("not reached")
 }
 
 func expectNoNotification(t *testing.T, ch <-chan *Notification) error {
@@ -34,7 +33,6 @@ func expectNoNotification(t *testing.T, ch <-chan *Notification) error {
 	case <-time.After(100 * time.Millisecond):
 		return nil
 	}
-	panic("not reached")
 }
 
 func expectEvent(t *testing.T, eventch <-chan ListenerEventType, et ListenerEventType) error {
@@ -47,7 +45,6 @@ func expectEvent(t *testing.T, eventch <-chan ListenerEventType, et ListenerEven
 	case <-time.After(1500 * time.Millisecond):
 		return fmt.Errorf("timeout")
 	}
-	panic("not reached")
 }
 
 func expectNoEvent(t *testing.T, eventch <-chan ListenerEventType) error {
@@ -57,7 +54,6 @@ func expectNoEvent(t *testing.T, eventch <-chan ListenerEventType) error {
 	case <-time.After(100 * time.Millisecond):
 		return nil
 	}
-	panic("not reached")
 }
 
 func newTestListenerConn(t *testing.T) (*ListenerConn, <-chan *Notification) {
@@ -219,8 +215,7 @@ func TestNotifyExtra(t *testing.T) {
 	defer db.Close()
 
 	if getServerVersion(t, db) < 90000 {
-		t.Log("skipping test due to old PG version")
-		return
+		t.Skip("skipping NOTIFY payload test since the server does not appear to support it")
 	}
 
 	l, channel := newTestListenerConn(t)
